Start on new lobby system.
Changes the protocol during the lobby phase to be more robust, in particular
ensuring slots are bound to user accounts.

Replaces the old network.h with a Rust binding to Tokio. Since C++ will have
to access QUIC in the future, for which the only viable library is quinn,
it makes sense to use Rust bindings for networking.