Prepare crate structure for new Rust server
We rewrote the client in Rust. The server is next.

Why? So that we can share code between the server and client, and because Rust is much
more enjoyable and robust to work with than C++.

The server code is relatively small compared to the client, so this isn't a massive undertaking.